- set configurable retries as default strategy in MessageFailureHandlerDefaultFactory - bug fix: fix type of storm configuration parameter for retry failure handler - bug fix: fix default offset returned by both failure strategies

+++ external/storm-kafka/src/jvm/storm/kafka/failures/MessageFailureHandlerDefaultFactory.java
@@ -41,10 +41,11 @@ public class MessageFailureHandlerDefaultFactory implements IMessageFailureHandl
         String key = "kafka.spout." + pm.getSpoutConfig().id + ".retries";
         Map stormConf = pm.getStormConfig();
         if (stormConf.containsKey(key)) {
-            retriesNumber = (Integer)stormConf.get(key);
+            //retriesNumber = Integer.parseInt((String)stormConf.get(key));
+            retriesNumber = ((Long)stormConf.get(key)).intValue();
         }
         else if (stormConf.containsKey("kafka.spout.retries")) {
-            retriesNumber = (Integer)stormConf.get("kafka.spout.retries");
+            retriesNumber = ((Long)stormConf.get("kafka.spout.retries")).intValue();
         }
 
         return retriesNumber;
